requirements.txt
astroid==2.4.2
colorama==0.4.3
isort==5.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
Pillow==7.2.0
pylint==2.6.0
rope==0.18.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
